pyproject.toml
[project]
name = "lidar-panoptic-segmentation"
version = "1.0.0"
description = "Modular, configuration-driven deep learning system for 3D panoptic segmentation and tree polygon extraction"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LiDAR Panoptic Segmentation Team"}
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "shapely>=2.0.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
full = [
    "pyarrow>=12.0.0",
    "hdbscan>=0.8.29",
    "laspy>=2.5.0",
    "plyfile>=0.9.0",
    "mlflow>=2.9.0",
    "wandb>=0.15.0",
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.14.0",
    "databricks-sdk>=0.12.0",
]

[tool.setuptools.packages.find]
include = ["lidar_panoptic_segmentation*"]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
